backend/requirements.txt
Flask
mysql-connector-python
python-dotenv
werkzeug
selenium

